Rename file and its children
